+++ tests/cli/commands/test_version.py
@@ -1,25 +1,22 @@
 import pystac
-from pystac.version import get_stac_version
-from stactools.cli.commands.version import create_version_command
-from stactools.core import __version__
+import stactools
+from click.testing import CliRunner
+from stactools.cli.cli import cli
 from stactools.core.utils.subprocess import call
-from stactools.testing import CliTestCase
 
 
-class VersionTest(CliTestCase):
-    def create_subcommand_functions(self):
-        return [create_version_command]
+def test_hello_world():
+    runner = CliRunner()
+    result = runner.invoke(cli, ["version"])
+    assert result.exit_code == 0
 
-    def test_hello_world(self):
-        result = self.run_command(["version"])
-        self.assertEqual(0, result.exit_code)
-        expected = (
-            f"stactools version {__version__}\n"
-            f"PySTAC version {pystac.__version__}\n"
-            f"STAC version {get_stac_version()}\n"
-        )
-        self.assertEqual(expected, result.output)
+    assert result.output == (
+        f"stactools version {stactools.core.__version__}\n"
+        f"PySTAC version {pystac.__version__}\n"
+        f"STAC version {pystac.version.get_stac_version()}\n"
+    )
 
-    def test_entry_point(self):
-        result = call(["stac", "version"])
-        self.assertEqual(0, result)
+
+def test_entry_point():
+    result = call(["stac", "version"])
+    assert result == 0